pyproject.toml
[project]
name = "fire-number"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "google-genai>=1.63.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "pandas>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.5",
    "pypdf>=6.7.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis[hiredis]>=7.2.0",
    "requests>=2.32.5",
    "pinecone>=6.0.0",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.40.0",
    "huggingface-hub>=1.11.0",
]
